openssl_like_aes256cbc_encrypt/requirement.txt
cffi==1.14.0
cryptography==2.9.2
pycparser==2.20
pycrypto==2.6.1
six==1.15.0
